Working on binary tree dungeon generation
Lots added to tree.c for generating the tree. Added convenience
functions to rect.c and vec2i.c for making rects and vecs. Added a
helper function to place tiles, accepting a position and a TILE_TYPE and
changing the tile on the global map variable to match - a MUCH needed
improvement. Also added a sweet orthogonal walk algorithm to place
corridors between two points.