Minor style changes
Playing around with different color choices for the dungeon.
draw_screen() now draws any explored tile that blocks movement or is a
stair, and any visible tile. Open/Close door functions in update.c just
replace the tile at the location with an open/closed door instead of
modifying the tile that was there.