Began work on map functions
Map is a doubly linked list, with each item containing an integer id
"lvl", and an array of tiles. Functions added for creation and
destruction of maps. Next step is to use the change level functions to
shift between the next/prev on the map... oh and update the drawing
functions to draw the current map... and the map assistance functions to
use the current map... Then, reading/writing the map from/to files.