SAVING AND LOADING WORKS!
Man - that was a challenge. The implementation is rough for now, and
just saves a "save_data.bin" file in the current directory. Load opens
this file and works BEAUTIFULLY. I'm sure everything would explode if
the save_data.bin file doesn't exist... Time to get the main menu up and
running I suppose!