Messages added!
Msg is a linked list, and the game will check for messages, pop them off
the list, draw them on the screen. Simple messages added for
opening/closing doors, going up/down stairs. Fixed an error in
mt_chance, it was returning the OPPOSITE of what it needed to... a 1/4
chance was incorrectly returned true 3/4 of the time (instead of 1/4).