Working on adding Monsters...
Player and enemies will be the same generic type of "Monster", so
player.h/player.c were renamed and some functions were modified to make
it more generic. Monsters will be kept in a linked list, and a pointer
to that list will be kept in Map. The player will be added/removed
to/from lists for whatever current level map they are on.

This update has MonsterList mostly complete, written as a circular
linked list.