Working on fixing the (many) errors with MList
MonsterList SHOULD work. It SHOULD be saving/loading properly. It isn't
though, so I've disabled map generation (player is on a big blank
field). This showed a bunch of errors in valgrind for trying to read
outside the map, so I added a simple in_bounds(x,y) error check to a
bunch of functions... and that cleared that up. Figure if I squash all
the errors until valgrind stops yelling, that the monsterlist should
work properly.

It's basically the same thing as the msglist... which has no (known)
issues. The struggle continues.