MList completely disabled, all other errors fixed
Zero errors, zero leaks, valgrind is happy everything is v nice. No idea
what's wrong with monsterlist, but I think I need to scrap it and try it
again (but better). If nothing else, I need to merge these changes with
master so at least the little errors are fixed.