Monsters now wake up and chase the player
Still working on the (many) kinks - monsters will attack other monsters
in their way, the game crashes if the player is killed, the time
management system is still being worked on, etc. Lots of things
commented out in the handle events/update routines that may need to be
scrapped entirely.