The player can now die!
Well, technically the player could already die, but then it would just
SEGFAULT and crash. Now, it displays a sad little message, waits for
input, and returns to the main screen. Also cleaned up main.c - opened
the file because it was dusty and realized there was a bunch of
commented out unnecessary stuff that should have been deleted ages ago.