Added new art drawing functions
Okay, so big code update temporarily - I didn't actually write 800 lines
of code. This is an intermediate step for a couple of ideas - one is
getting rid of being dependent on curses. I don't like how the
title/help/new player/etc screens were drawn by calling curses calls
instead of using the abstracted screen calls the game uses. Getting rid
of the "dependency" of curses means that in the future, I can change who
is actually drawing on the screen easily (SDL, for instance). The other
thing that's accomplished here is the little script interpreter that
wrote those 800 lines for me will be added in to art.c, and the art will
be read from config files to generate the art "on the fly" soon.