Minor fix to split_string in slist.c
While working on reading in the art files, I found a silly bug that
would cause a segfault... if a string ended in the delimeter specified
in split string, it would try and assign a null character to the end of
a string that didn't exist. Quick fix once I figured it out.