"Finished" reading in the art config file
It works! Little wonky with extended characters (127-255), so I have to
figure that out eventually. Not entirely sure all the effort was worth
it, except as an exercise in actually being able to do it. Reading in an
external file **looks** a heck of a lot cleaner in code than hard coding
all the art (art.c is like, a thousand lines shorter). Leaving it as is
for now to work on other things.