Made a fancy generic linked list!
Obviously put it in the toolbox, because it's super generic and can be
ued for all sorts of things. It will (eventually) replace things like
Vec2iList and RectList (and MList and SList and Msg) - no reason for
each of those to have a dedicated list type now that this bad boy
(almosts) exists. Still need to work on going from single link to double
link, and adding some other convience functions.