Monsters can now have different speeds!
HOLY COW. This was a doozy of a logic update and is going to take a bit
of tidying to cleanup. Everything works here, monsters gain energy each
turn (currently 25 for testing) of the clock. When they have enough
energy they use it for their turn. A monster with speed 50 will act
twice as often as a monster with speed 100, and four times as often as a
monster with speed 200.