Working on rewriting all drawing functions
Everything should draw to a Glyph* "screen buffer", and not directly to
the screen... Started with draw_game() in draw.c. Monsters temporarily
disabled by commenting out the add monsters part of dungeon generation.
Map width/height fiddled with. Art drawing code in art.c fixed so that
screen width/height are constant. Lots more to do, but it's a start.