Fixed New Player screen error checking
This also included fixing the way keyboard input is handled in
term_engine.c/h. Single character requests from the keyboard are
currently set up to be "blocking" - it waits for read to get a character
that isn't '\0'. String requests from the terminal use fgets() and
returns NULL if an empty string is entered. On the new player screen, if
the user hits return without entering a name, the game pops up an error
message and makes the user try again.