Random name functions added!
Couple of fun new things here. Added a function to draw.c to draw "menu"
messages, basically a prompt with a list of options that returns a
character tied to an option. Added in the files from the Markov project,
and updated files from the toolbox. Finally, added a function to
generate a list of random names for the player, presented with a menu,
during the new player screen.